* test now renders three textured squares, one controlled by the mouse, with support for vsync enable/disable, and double buffering enable/disable

+++ src/Mouse.cpp
@@ -66,8 +66,8 @@ struct MouseState Mouse::readDevice()
 		m_mouse_cooldown--;
 	}
 
-	s.x = (int)(xpos * Scale);
-	s.y = (int)(ypos * Scale);
+	s.x = (xpos * Scale);
+	s.y = (ypos * Scale);
 
 
 	return s;
